# Revert "Migrate to aioredis, return Cache-Control header even for cached response"