--- a/requirements/prod.txt
+++ b/requirements/prod.txt
@@ -1,7 +1,6 @@
 aioboto3==9.0.0
 aiohttp==3.7.4.post0
 aiohttp-jinja2==1.4.2
-aioredis==2.0.0b1
 APScheduler==3.7.0
 asgiref==3.3.4
 async-timeout==3.0.1
@@ -62,6 +61,7 @@ python-dateutil==2.6.0
 python-dotenv==0.17.1
 pytz==2021.1
 PyYAML==5.4.1
+redis==3.5.3
 requests==2.25.1
 requests-cache==0.6.4
 streamscrobbler @ git+https://github.com/Horrendus/streamscrobbler-python@ec8fd4e1549b7cba25833950aa565b17700453e0