--- a/requirements/test.txt
+++ b/requirements/test.txt
@@ -1,4 +1,3 @@
-asynctest==0.13.0
 coverage>=4.0.3
 pluggy>=0.3.1
 py>=1.4.31