--- a/src/api/__init__.py
+++ b/src/api/__init__.py
@@ -1,7 +1,5 @@
 import connexion
 
-import aioredis
-
 from api import encoder
 from api.tracing import configure_tracer
 from api.cors import install_cors_middleware
@@ -16,16 +14,3 @@ def create_app():
 
     configure_tracer(app)
     return app
-
-
-async def create_redis_pool(app):
-    """Create Redis connection pool"""
-    redis_url = app["redis_url"]
-    pool = await aioredis.from_url(redis_url)
-    app["redis_pool"] = pool
-
-
-async def close_redis_pool(app):
-    """Close Redis connection pool"""
-    pool = app["redis_client"]
-    await pool.close()