Reverts jjst/now-playing#120

This change is causing performance regressions: p50 observed on Lightstep went from ~5ms to ~2s. p95 for *cached responses* from 5-10ms to 11 seconds.